+++ .gitignore
@@ -1,9 +1,10 @@
 # dependenices
 node_modules
-src/node_modules/@sapper
 
-# build
-__sapper__
+# production build
+.svelte
+build
+functions
 
 # deploy config
 .netlify
