+++ lighthouserc.yml
@@ -1,7 +1,7 @@
-# https://github.com/GoogleChrome/lighthouse-ci/blob/master/docs/configuration.md
+# https://github.com/GoogleChrome/lighthouse-ci/blob/main/docs/configuration.md
 ci:
   collect:
-    staticDistDir: ./__sapper__/export
+    staticDistDir: ./build
   assert:
     presets: lighthouse:no-pwa
     assertions:
