+++ package.json
@@ -2,61 +2,46 @@
   "private": true,
   "name": "afara.foundation",
   "title": "Afara e.V.",
-  "description": "Svelte + Contentful-powered homepage of the student-run non-profit Afara e.V. registered in Würzburg, Germany",
+  "description": "SvelteKit + Contentful-powered homepage of the student-run non-profit initiative Afara from Würzburg, Germany supporting child care projects in Africa.",
   "homepage": "https://afara.foundation",
   "version": "1.0.0",
+  "type": "module",
   "author": "Janosh Riebesell <janosh.riebesell@gmail.com>",
-  "repository": "https://github.com/janosh/afara-svelte",
+  "repository": "https://github.com/janosh/afara",
   "scripts": {
-    "dev": "sapper dev",
-    "build": "sapper build",
-    "export": "sapper export",
-    "start": "sapper dev",
-    "serve": "sapper export && sirv __sapper__/export"
-  },
-  "dependencies": {
-    "algoliasearch": "^4.8.5",
-    "compression": "^1.7.1"
+    "dev": "svelte-kit dev",
+    "disabled-prebuild": "node src/utils/algolia.js",
+    "build": "svelte-kit build",
+    "preview": "svelte-kit preview",
+    "serve": "yarn build && yarn preview"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.8",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.13.9",
-    "@babel/preset-env": "^7.13.9",
-    "@babel/runtime": "^7.13.9",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.2.0",
-    "@rollup/plugin-replace": "^2.4.1",
-    "@svg-icons/bootstrap": "^1.40.0",
-    "@svg-icons/boxicons-regular": "^1.39.1",
-    "@svg-icons/boxicons-solid": "^1.39.1",
-    "@svg-icons/entypo": "^1.39.1",
-    "@svg-icons/fa-brands": "^1.39.1",
-    "@svg-icons/fa-solid": "^1.39.1",
-    "@svg-icons/heroicons-solid": "^1.39.1",
-    "@svg-icons/ionicons-solid": "^1.39.1",
-    "@svg-icons/material-sharp": "^1.39.1",
-    "@svg-icons/octicons": "^1.39.1",
-    "@svg-icons/remix-fill": "^1.39.1",
-    "@svg-icons/simple-icons": "^1.40.0",
-    "contentful-management": "^7.8.0",
-    "cross-fetch": "^3.0.6",
+    "@rollup/plugin-replace": "^2.4.2",
+    "@sveltejs/adapter-netlify": "^1.0.0-next.10",
+    "@sveltejs/kit": "^1.0.0-next.92",
+    "@svicons/bootstrap": "^0.1.3",
+    "@svicons/boxicons-regular": "^0.1.3",
+    "@svicons/boxicons-solid": "^0.1.3",
+    "@svicons/entypo": "^0.1.3",
+    "@svicons/fa-brands": "^0.1.3",
+    "@svicons/fa-solid": "^0.1.3",
+    "@svicons/heroicons-solid": "^0.1.3",
+    "@svicons/ionicons-solid": "^0.1.3",
+    "@svicons/material-sharp": "^0.1.3",
+    "@svicons/octicons": "^0.1.3",
+    "@svicons/remix-fill": "^0.1.3",
+    "@svicons/simple-icons": "^0.1.3",
+    "contentful-management": "^7.17.1",
     "dotenv": "^8.2.0",
-    "eslint": "^7.21.0",
-    "eslint-plugin-svelte3": "^3.1.2",
-    "js-yaml": "^4.0.0",
-    "marked": "^2.0.1",
-    "polka": "^0.5.2",
+    "eslint": "^7.25.0",
+    "eslint-plugin-svelte3": "^3.2.0",
+    "js-yaml": "^4.1.0",
+    "marked": "^2.0.3",
     "prettier": "^2.2.1",
-    "prettier-plugin-svelte": "^2.1.6",
-    "rollup": "^2.40.0",
-    "rollup-plugin-svelte": "^7.1.0",
+    "prettier-plugin-svelte": "^2.2.0",
     "rollup-plugin-svelte-svg": "^0.2.3",
-    "rollup-plugin-terser": "^7.0.2",
-    "sapper": "^0.29.1",
-    "sirv": "^1.0.11",
-    "sirv-cli": "^1.0.11",
-    "svelte": "^3.35.0",
-    "svelte-algolia": "^0.1.5"
+    "svelte": "^3.37.0",
+    "svelte-algolia": "^0.1.6",
+    "vite": "^2.2.3"
   }
 }
