+++ src/app.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="de">
+  <head>
+    <title>AFARA e.V.</title>
+    <meta name="author" content="Janosh Riebesell" />
+    <meta
+      name="description"
+      content="Afara ist ein ehrenamtlicher studentischer Verein aus Würzburg der Bildungsprojekte in Afrika fördert."
+    />
+    <meta charset="utf-8" />
+    <link rel="stylesheet" href="global.css" />
+    <link rel="icon" href="/favicon.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <link rel="manifest" href="manifest.json" crossorigin="use-credentials" />
+
+    %svelte.head%
+  </head>
+
+  <body>
+    <div id="svelte">%svelte.body%</div>
+  </body>
+</html>
