--- src/client.js
@@ -1,3 +1,0 @@
-import { start } from '@sapper/app'
-
-start({ target: document.body })
