+++ src/components/MapEmbed.svelte
@@ -1,8 +1,8 @@
 <script>
-  import { stores } from '@sapper/app'
+  import { session } from '$app/stores'
+
   export let place, title
 
-  const { session } = stores()
   const { GOOGLE_MAPS_API_KEY: key } = $session
 
   const src = `https://google.com/maps/embed/v1/place?q=${place}&key=${key}`
