+++ src/components/Search.svelte
@@ -1,20 +1,19 @@
 <script>
   import { onMount } from 'svelte'
-  import algoliasearch from 'algoliasearch/lite'
-  import { stores } from '@sapper/app'
-  import Search from '@svg-icons/fa-solid/search.svg'
+  import { session } from '$app/stores'
+
+  import Search from '@svicons/fa-solid/search.svelte'
 
   import SearchHit from './SearchHit.svelte'
   import { onClickOutside } from '../utils/actions'
 
-  const { session } = stores()
   const { ALGOLIA_APP_ID: appId, ALGOLIA_SEARCH_KEY: searchKey } = $session
 
   export let indices = []
   let client, input, query, promise
   let hasFocus = false
 
-  onMount(() => (client = algoliasearch(appId, searchKey)))
+  onMount(() => (client = window.algoliasearch(appId, searchKey)))
 
   const processHits = (hits) =>
     hits.map((hit) => {
@@ -33,7 +32,12 @@
   }
 
   const style = `vertical-align: text-bottom; z-index: 0;`
+  const src = `https://cdn.jsdelivr.net/npm/algoliasearch@latest/dist/algoliasearch-lite.umd.js`
 </script>
+
+<svelte:head>
+  <script async defer {src}></script>
+</svelte:head>
 
 <aside use:onClickOutside={() => (hasFocus = false)}>
   <input
