+++ src/routes/[...slug].svelte
@@ -1,9 +1,10 @@
 <script context="module">
   import { fetchPage } from '../utils/queries'
 
-  export async function preload({ params }) {
-    const page = await fetchPage(params.slug.join(`/`))
-    return { page }
+  export async function load({ page: { params } }) {
+    const page = await fetchPage(params.slug)
+
+    return { props: { page } }
   }
 </script>
