--- src/template.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="de">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width,initial-scale=1.0" />
-    <meta name="theme-color" content="#333333" />
-    <meta
-      name="description"
-      content="Afara e.V. ist ein ehrenamtlicher studentischer Verein aus Würzburg der Bildungsprojekte in Afrika unterstützt."
-    />
-
-    %sapper.base%
-
-    <link rel="stylesheet" href="global.css" />
-    <link rel="manifest" href="manifest.json" crossorigin="use-credentials" />
-    <link rel="icon" type="image/svg" href="favicon.svg" />
-
-    <!-- Sapper creates a <script> tag containing `src/client.js`
-      and anything else it needs to hydrate the app and
-      initialize the router -->
-    %sapper.scripts%
-
-    <!-- Sapper generates a <style> tag containing critical CSS
-      for the current page. CSS for the rest of the app is
-      lazily loaded when it precaches secondary pages -->
-    %sapper.styles%
-
-    <!-- Contents of the <svelte:head> component appear here, if the current page has one -->
-    %sapper.head%
-  </head>
-  <!-- The app will render inside body because `src/client.js` references it -->
-  <body>
-    %sapper.html%
-  </body>
-</html>
