+++ src/utils/queries.js
@@ -1,69 +1,13 @@
 /* eslint-disable indent */
-import 'cross-fetch/polyfill'
-import marked from 'marked'
+import marked from './marked'
 import yaml from 'js-yaml'
-
-const renderer = {
-  // responsive markdown images
-  image(href, title, text) {
-    if (href?.includes(`images.ctfassets.net`) && !href.endsWith(`.svg`)) {
-      title = title ? `title="${title}"` : ``
-
-      const srcSet = (params) =>
-        [900, 600, 400]
-          .map((width) => `${href}?w=${width}&${params} ${width}w`)
-          .join(`, `)
-
-      return `
-      <picture>
-        <source srcset="${srcSet(`q=80&fit=fill&fm=webp`)}" type="image/webp" />
-        <source srcset="${srcSet(`q=80&fit=fill`)}" />
-        <img src="${href}?w=900&q=80" alt="${text}" ${title} loading="lazy" />
-      </picture>`
-    }
-
-    return false // delegate to default marked image renderer
-  },
-  // add Sapper prefetching for local markdown links
-  link(href, title, text) {
-    if (!href.startsWith(`http`) && !href.startsWith(`www`)) {
-      title = title ? `title="${title}"` : ``
-      return `<a sapper:prefetch href="${href}" ${title}>${text}</a>`
-    }
-    return false // delegate to default marked link renderer
-  },
-  // responsive iframes for video embeds
-  codespan(code) {
-    if (code.startsWith(`youtube:`) || code.startsWith(`vimeo:`)) {
-      const [platform, id] = code.split(/:\s?/)
-      const embed = {
-        youtube: (id) => `https://youtube.com/embed/${id}`,
-        vimeo: (id) => `https://player.vimeo.com/video/${id}`,
-      }
-      // padding-top: 56.25%; corresponds to 16/9 = most common video aspect ratio
-      return `
-        <div style="padding-top: 56.25%; position: relative;">
-          <iframe
-            title="${platform} video"
-            loading="lazy"
-            src="${embed[platform](id)}"
-            style="position: absolute; top: 0; left: 0; width: 100%; height: 100%; border: 0;"
-            allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture;"
-            allowfullscreen></iframe>
-        </div>`
-    }
-    return false // delegate to default marked codespan renderer
-  },
-}
-
-marked.use({ renderer })
 
 const prefixSlug = (prefix) => (obj) => {
   obj.slug = prefix + obj.slug
   return obj
 }
 
-export async function ctfFetch(query) {
+export async function contentfulFetch(query) {
   const token = process.env.CONTENTFUL_ACCESS_TOKEN
   const id = process.env.CONTENTFUL_SPACE_ID
 
@@ -149,7 +93,7 @@ const pagesQuery = `{
 
 export async function fetchPage(slug) {
   if (!slug) throw `fetchPage requires a slug, got '${slug}'`
-  const data = await ctfFetch(pageQuery(slug))
+  const data = await contentfulFetch(pageQuery(slug))
   const page = data?.pages?.items[0]
   if (page?.yaml) {
     page.yaml = yaml.load(page.yaml)
@@ -160,7 +104,7 @@ export async function fetchPage(slug) {
 }
 
 export async function fetchPages() {
-  const data = await ctfFetch(pagesQuery)
+  const data = await contentfulFetch(pagesQuery)
   return data?.pages?.items?.map(renderBody)
 }
 
@@ -206,13 +150,13 @@ async function processPost(post) {
 
 export async function fetchPost(slug) {
   if (!slug) throw `fetchPost requires a slug, got '${slug}'`
-  const data = await ctfFetch(postQuery(slug))
+  const data = await contentfulFetch(postQuery(slug))
   const post = data?.posts?.items[0]
   return processPost(post)
 }
 
 export async function fetchPosts() {
-  const data = await ctfFetch(postsQuery)
+  const data = await contentfulFetch(postsQuery)
   const posts = data?.posts?.items
   return await Promise.all(posts.map(processPost))
 }
@@ -227,6 +171,6 @@ const yamlQuery = (title) => `{
 
 export async function fetchYaml(title) {
   if (!title) throw `fetchYaml requires a title, got '${title}'`
-  const { yml } = await ctfFetch(yamlQuery(title))
+  const { yml } = await contentfulFetch(yamlQuery(title))
   return yaml.load(yml?.items[0]?.data)
 }
